stix_shifter_modules/azure_log_analytics/requirements.txt
azure-monitor-query==1.0.2
pandas==1.5.2
jsonref==1.1.0